moved dev_team folder into multi_agents folder & edited imports - discord bot query runs to completion
