--- dev_team/agents/__init__.py
@@ -1,8 +1,0 @@
-from .github_agent import GithubAgent
-from .repo_analyzer_agent import RepoAnalyzerAgent
-from .web_search_agent import WebSearchAgent
-from .rubber_ducker_agent import RubberDuckerAgent
-from .tech_lead_agent import TechLeadAgent
-from .file_search_agent import FileSearchAgent
-
-__all__ = ['GithubAgent', 'RepoAnalyzerAgent', 'WebSearchAgent', 'FileSearchAgent', 'RubberDuckerAgent', 'TechLeadAgent']
