--- dev_team/requirements.txt
@@ -1,11 +1,0 @@
-PyGithub
-langchain
-
-openai
-colorama
-langgraph
-gpt_researcher
-python-dotenv
-
-langchain-openai
-faiss-cpu
