+++ multi_agents/dev_team/agents/file_search_agent.py
@@ -1,5 +1,4 @@
 from .utils.llms import call_model
-from dev_team.agents import GithubAgent
 import os
 sample_json = """
 {
