+++ multi_agents/requirements.txt
@@ -5,3 +5,9 @@ python-dotenv
 weasyprint
 json5
 loguru
+PyGithub
+langchain
+openai
+colorama
+langchain-openai
+faiss-cpu
